requirements.txt
opencv-python
face_recognition
gpiozero
smbus2